+++ CHANGELOG.md
@@ -6,6 +6,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 ## [Unreleased]
 
+## [2.0.1] - 2026-09-13
+
+### Fixed
+
+- Quiet `aioesphomeapi` / `zeroconf` DEBUG packet dumps that flooded
+  `debug.log`; **Debug + API Stream** uses short one-line `API state` messages.
+
 ## [2.0.0] - 2026-08-16
 
 ### Changed